requirements.txt
google-generativeai==0.7.2
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1 
